+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - switched to `kasperl` library for base API and generic pipeline plugins
 - added `thinning` filter for skeletonizing binary images, based on
   Lingdong Huang's code at https://github.com/LingDong-/skeleton-tracing/blob/master/py/trace_skeleton.py
+- the `find-contours` filter now has optional min/max size restrictions for the objects
 
 
 0.0.10 (2025-07-11)